fix: update investments layout
